+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Added
+- Support for theme-park.dev themes.
+
+### Fixed
+- Wrong css when reapplying the default theme.
+
 ## [3.4.1] - 2021-08-11
 ### Added
 - Toggle to disable embeds.